第一次上传

+++ NameCube/Setting/Document.xaml.cs
@@ -37,21 +37,39 @@ namespace NameCube.Setting
         private void CardAction_Click_1(object sender, RoutedEventArgs e)
         {
             DialogResult result = System.Windows.Forms.MessageBox.Show("确定删除日志及备份文件吗？", "确认", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (result == DialogResult.Yes) 
+            if (result == DialogResult.Yes)
             {
-                Directory.Delete(System.IO.Path.Combine(GlobalVariables.configDir, "logs"),true);
-                Directory.Delete(System.IO.Path.Combine(GlobalVariables.configDir, "Mode_data", "MemoryFactoryMode", "Backups"),true);
-                System.Windows.MessageBox.Show("删除成功");
+                try
+                {
+                    if (File.Exists(System.IO.Path.Combine(GlobalVariables.configDir, "logs")))
+                        Directory.Delete(System.IO.Path.Combine(GlobalVariables.configDir, "logs"), true);
+                    if (File.Exists(System.IO.Path.Combine(GlobalVariables.configDir, "Mode_data", "MemoryFactoryMode", "Backups")))
+                        Directory.Delete(System.IO.Path.Combine(GlobalVariables.configDir, "Mode_data", "MemoryFactoryMode", "Backups"), true);
+                    System.Windows.MessageBox.Show("删除成功");
+                }
+                catch (Exception ex)
+                {
+                    System.Windows.MessageBox.Show(ex.Message);
+                }
             }
+
         }
 
         private void CardAction_Click_2(object sender, RoutedEventArgs e)
         {
             DialogResult result = System.Windows.Forms.MessageBox.Show("你确定继续执行操作吗？这将从你的硬盘上彻底删除配置文件", "确认", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-            if (result == DialogResult.Yes) { 
-                Directory.Delete(GlobalVariables.configDir, true);
-                System.Windows.MessageBox.Show("删除成功，请自行启动软件");
-                System.Windows.Application.Current.Shutdown();
+            if (result == DialogResult.Yes) {
+                try
+                {
+                    Directory.Delete(GlobalVariables.configDir, true);
+                    System.Windows.MessageBox.Show("删除成功，请自行启动软件");
+                    System.Windows.Application.Current.Shutdown();
+                }
+                catch (Exception ex)
+                {
+                    System.Windows.MessageBox.Show(ex.Message);
+                }
+
             }
         }
 
@@ -88,11 +106,19 @@ namespace NameCube.Setting
                 openFileDialog.Filter = "压缩包 (*.zip)|*.zip|所有文件 (*.*)|*.*";
                 if (openFileDialog.ShowDialog() == DialogResult.OK)
                 {
-                    Directory.Delete(GlobalVariables.configDir, true);
-                    Directory.CreateDirectory(GlobalVariables.configDir);
-                    SevenZipCompressor.Decompress(openFileDialog.FileName, GlobalVariables.configDir);
-                    System.Windows.MessageBox.Show("覆盖成功，请自行启动软件");
-                    System.Windows.Application.Current.Shutdown();
+                    try
+                    {
+                        Directory.Delete(GlobalVariables.configDir, true);
+                        Directory.CreateDirectory(GlobalVariables.configDir);
+                        SevenZipCompressor.Decompress(openFileDialog.FileName, GlobalVariables.configDir);
+                        System.Windows.MessageBox.Show("覆盖成功，请自行启动软件");
+                        System.Windows.Application.Current.Shutdown();
+                    }
+                    catch (Exception ex)
+                    {
+                        System.Windows.MessageBox.Show(ex.Message);
+                    }
+                    
                 }
                 
             }
